runner: move port-forwarding code under pkg/
Move the port-forwarding code from cmd/lvh in pkg/runner. This allows
using this code from other places (namely, the tetragon tester which is
where most of the runner code comes from). We should do this for other
parts of the runner so that cmd/lvh and tetragon tester can share the
same code.
